Get Main component testing working
