+++ client/components/Main/Main.test.js
@@ -1,52 +1,8 @@
 import React from 'react';
-import { shallow } from '@jest-helpers/renderer';
+import { mount } from 'enzyme';
 import { MockSimpleChild } from '@jest-helpers/components';
-import { Main as Component } from './Main';
-
-const mockDeck = [
-  {
-    cardIndex: 0,
-    hasMatched: false,
-    image: 'images/svg-cards/2_of_hearts.svg',
-    isFlipped: false,
-    name: '2 of hearts',
-  },
-  {
-    cardIndex: 1,
-    hasMatched: false,
-    image: 'images/svg-cards/3_of_hearts.svg',
-    isFlipped: false,
-    name: '3 of hearts',
-  },
-  {
-    cardIndex: 2,
-    hasMatched: false,
-    image: 'images/svg-cards/4_of_hearts.svg',
-    isFlipped: false,
-    name: '4 of hearts',
-  },
-  {
-    cardIndex: 0,
-    hasMatched: false,
-    image: 'images/svg-cards/2_of_hearts.svg',
-    isFlipped: false,
-    name: '2 of hearts',
-  },
-  {
-    cardIndex: 1,
-    hasMatched: false,
-    image: 'images/svg-cards/3_of_hearts.svg',
-    isFlipped: false,
-    name: '3 of hearts',
-  },
-  {
-    cardIndex: 2,
-    hasMatched: false,
-    image: 'images/svg-cards/4_of_hearts.svg',
-    isFlipped: false,
-    name: '4 of hearts',
-  },
-];
+import mockDeck from '../../../testHelpers/__mocks__/mockDeck';
+import Component from './Main';
 
 const mockProps = {
   game: {
@@ -57,15 +13,14 @@ const mockProps = {
   },
 };
 
-
 describe(`<Main />`, () => {
   test(`It doesn't blow up` , () => {
-    const tree = shallow(<Component {...mockProps}><MockSimpleChild /></Component>);
+    const tree = mount(<Component {...mockProps}><MockSimpleChild /></Component>);
     expect(tree.length).toEqual(1);
   });
 
-  // test(`Renders correctly`, () => {
-  //   const main = mount(<Component {...mockProps} />);
-  //   expect(main.debug()).toMatchSnapshot();
-  // });
+  test(`Renders correctly`, () => {
+    const main = mount(<Component {...mockProps}><MockSimpleChild /></Component>);
+    expect(main.debug()).toMatchSnapshot();
+  });
 });
