+++ testHelpers/__mocks__/mockDeck.js
@@ -1,0 +1,46 @@
+const mockDeck = [
+  {
+    cardIndex: 0,
+    hasMatched: false,
+    image: 'images/svg-cards/2_of_hearts.svg',
+    isFlipped: false,
+    name: '2 of hearts',
+  },
+  {
+    cardIndex: 1,
+    hasMatched: false,
+    image: 'images/svg-cards/3_of_hearts.svg',
+    isFlipped: false,
+    name: '3 of hearts',
+  },
+  {
+    cardIndex: 2,
+    hasMatched: false,
+    image: 'images/svg-cards/4_of_hearts.svg',
+    isFlipped: false,
+    name: '4 of hearts',
+  },
+  {
+    cardIndex: 0,
+    hasMatched: false,
+    image: 'images/svg-cards/2_of_hearts.svg',
+    isFlipped: false,
+    name: '2 of hearts',
+  },
+  {
+    cardIndex: 1,
+    hasMatched: false,
+    image: 'images/svg-cards/3_of_hearts.svg',
+    isFlipped: false,
+    name: '3 of hearts',
+  },
+  {
+    cardIndex: 2,
+    hasMatched: false,
+    image: 'images/svg-cards/4_of_hearts.svg',
+    isFlipped: false,
+    name: '4 of hearts',
+  },
+];
+
+export default mockDeck;
